refactor(coding-agent): extract gh shared module
Split the shared type vocabulary and universal helpers out of the 3.7k-line gh.ts into a leaf gh-shared module; gh.ts keeps re-exporting the public surface unchanged.

Op: compress
