+++ packages/coding-agent/src/tools/gh-shared.ts
@@ -1,0 +1,234 @@
+import * as path from "node:path";
+import type { AgentToolResult } from "@oh-my-pi/pi-agent-core";
+import { untilAborted } from "@oh-my-pi/pi-utils";
+import { type } from "arktype";
+import * as git from "../utils/git";
+import type { OutputMeta } from "./output-meta";
+import { ToolError } from "./tool-errors";
+import { toolResult } from "./tool-result";
+
+export const githubSchema = type({
+	op: type(
+		"'repo_view' | 'pr_create' | 'pr_checkout' | 'pr_push' | 'search_issues' | 'search_prs' | 'search_code' | 'search_commits' | 'search_repos' | 'run_watch'",
+	).describe("github operation"),
+	"repo?": type("string").describe("owner/repo"),
+	"branch?": type("string").describe("branch"),
+	"pr?": type("string | string[]").describe("pr number, url, or branch"),
+	"force?": type("boolean").describe("reset existing local branch"),
+	"forceWithLease?": type("boolean").describe("force-with-lease push"),
+	"title?": type("string").describe("pr title"),
+	"body?": type("string").describe("pr body markdown"),
+	"base?": type("string").describe("pr base branch"),
+	"head?": type("string").describe("pr head branch"),
+	"draft?": type("boolean").describe("open pr as draft"),
+	"fill?": type("boolean").describe("auto-fill pr title/body from commits"),
+	"reviewer?": type("string[]").describe("reviewers"),
+	"assignee?": type("string[]").describe("assignees"),
+	"label?": type("string[]").describe("labels"),
+	"query?": type("string").describe("search query"),
+	"since?": type("string").describe("lower-bound date filter"),
+	"until?": type("string").describe("upper-bound date filter"),
+	"dateField?": type("'created' | 'updated'").describe("date field"),
+	"limit?": type("number").describe("max results"),
+	"run?": type("string").describe("actions run id or url"),
+	"tail?": type("number").describe("log lines per failed job"),
+});
+
+export type GithubInput = typeof githubSchema.infer;
+
+export interface GhToolDetails {
+	meta?: OutputMeta;
+	artifactId?: string;
+	repo?: string;
+	branch?: string;
+	worktreePath?: string;
+	remote?: string;
+	remoteBranch?: string;
+	headSha?: string;
+	runId?: number;
+	runIds?: number[];
+	status?: string;
+	conclusion?: string;
+	failedJobs?: string[];
+	watch?: GhRunWatchViewDetails;
+	checkouts?: GhPrCheckoutSummary[];
+}
+
+export interface GhPrCheckoutSummary {
+	prNumber?: number;
+	url?: string;
+	branch: string;
+	worktreePath: string;
+	remote: string;
+	remoteBranch: string;
+	reused: boolean;
+}
+
+export interface GhRunWatchJobDetails {
+	id: number;
+	name: string;
+	status?: string;
+	conclusion?: string;
+	durationSeconds?: number;
+	url?: string;
+}
+
+export interface GhRunWatchRunDetails {
+	id: number;
+	workflowName?: string;
+	displayTitle?: string;
+	status?: string;
+	conclusion?: string;
+	branch?: string;
+	headSha?: string;
+	url?: string;
+	jobs: GhRunWatchJobDetails[];
+}
+
+export interface GhRunWatchFailedLogDetails {
+	runId: number;
+	workflowName?: string;
+	jobName: string;
+	conclusion?: string;
+	tail?: string;
+	available: boolean;
+}
+
+export interface GhRunWatchViewDetails {
+	mode: "run" | "commit";
+	state: "watching" | "completed";
+	repo: string;
+	branch?: string;
+	headSha?: string;
+	pollCount?: number;
+	note?: string;
+	run?: GhRunWatchRunDetails;
+	runs?: GhRunWatchRunDetails[];
+	failedLogs?: GhRunWatchFailedLogDetails[];
+}
+
+export interface GhUser {
+	login?: string;
+	name?: string | null;
+}
+
+export interface GhLabel {
+	name?: string;
+}
+
+export function normalizeText(value: string | null | undefined): string {
+	return (value ?? "").replaceAll("\r\n", "\n").replaceAll("\r", "\n").replaceAll("\t", "    ").trim();
+}
+
+export function normalizeOptionalString(value: string | null | undefined): string | undefined {
+	const normalized = value?.trim();
+	return normalized ? normalized : undefined;
+}
+
+export function requireNonEmpty(value: string | null | undefined, label: string): string {
+	const normalized = normalizeOptionalString(value);
+	if (!normalized) {
+		throw new ToolError(`${label} must not be empty`);
+	}
+	return normalized;
+}
+
+export async function requireCurrentGitBranch(cwd: string, signal?: AbortSignal): Promise<string> {
+	const branch = await git.branch.current(cwd, signal);
+	if (!branch) {
+		throw new ToolError("Current git branch is unavailable. Pass `branch` or `run` explicitly.");
+	}
+
+	return branch;
+}
+
+export function formatAuthor(author: GhUser | null | undefined): string | undefined {
+	if (!author) return undefined;
+	if (author.login) return `@${author.login}`;
+	if (author.name) return author.name;
+	return undefined;
+}
+
+export function formatLabels(labels: GhLabel[] | undefined): string | undefined {
+	const names = labels?.map(label => label.name).filter((value): value is string => Boolean(value)) ?? [];
+	if (names.length === 0) return undefined;
+	return names.join(", ");
+}
+
+export function pushLine(lines: string[], label: string, value: string | number | boolean | undefined): void {
+	if (value === undefined || value === "") return;
+	lines.push(`${label}: ${value}`);
+}
+
+/**
+ * Process-lifetime cache of `gh repo view --json nameWithOwner` lookups keyed
+ * by absolute cwd. Avoids repeated `gh` chatter when the same protocol handler
+ * or tool call resolves the default repo many times in a row.
+ *
+ * The shared lookup is intentionally **not** bound to any caller's
+ * AbortSignal. Cancelling one caller would otherwise kill the underlying
+ * `gh repo view` for every concurrent waiter on the same cwd. Each caller's
+ * signal is honored at the wait point via `untilAborted` instead, so an abort
+ * unwinds only that caller.
+ */
+const DEFAULT_REPO_RESOLVED = new Map<string, string>();
+
+const DEFAULT_REPO_INFLIGHT = new Map<string, Promise<string>>();
+
+export async function resolveDefaultRepoMemoized(cwd: string, signal?: AbortSignal): Promise<string> {
+	const key = path.resolve(cwd);
+	const ready = DEFAULT_REPO_RESOLVED.get(key);
+	if (ready) return ready;
+	let pending = DEFAULT_REPO_INFLIGHT.get(key);
+	if (!pending) {
+		pending = (async () => {
+			// No caller signal: this lookup is shared across every concurrent
+			// waiter on the same cwd.
+			const resolved = await git.github.text(cwd, [
+				"repo",
+				"view",
+				"--json",
+				"nameWithOwner",
+				"-q",
+				".nameWithOwner",
+			]);
+			const value = requireNonEmpty(resolved, "repo");
+			DEFAULT_REPO_RESOLVED.set(key, value);
+			return value;
+		})();
+		// Drop the in-flight slot on settle so failures don't poison the cache
+		// and so a successful resolution survives only in `DEFAULT_REPO_RESOLVED`.
+		void pending.then(
+			() => DEFAULT_REPO_INFLIGHT.delete(key),
+			() => DEFAULT_REPO_INFLIGHT.delete(key),
+		);
+		DEFAULT_REPO_INFLIGHT.set(key, pending);
+	}
+	return untilAborted(signal, pending);
+}
+
+function appendArtifactReference(text: string, artifactId: string | undefined, label: string): string {
+	if (!artifactId) {
+		return text;
+	}
+
+	return `${text}\n\n${label}: artifact://${artifactId}`;
+}
+
+export function buildTextResult(
+	text: string,
+	sourceUrl?: string,
+	details?: GhToolDetails,
+	options?: { artifactId?: string; artifactLabel?: string; useless?: boolean },
+): AgentToolResult<GhToolDetails> {
+	const builder = toolResult<GhToolDetails>(details).text(
+		appendArtifactReference(text, options?.artifactId, options?.artifactLabel ?? "Saved artifact"),
+	);
+	if (sourceUrl) {
+		builder.sourceUrl(sourceUrl);
+	}
+	if (options?.useless) {
+		builder.useless();
+	}
+	return builder.done();
+}
